+++ src/addons/vamiga/Emulator/CPU/Moira/MoiraExceptions.h
@@ -31,4 +31,4 @@ void execTrapException(int nr);
 void execPrivilegeException();
 
 // Emulates an interrupt exception
-void execIrqException(u8 level);
+void execIrqException(int level);
